chore(create-email): start changing structure for create-email
